requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
